+++ aws/build.sh
@@ -2,7 +2,11 @@
 set -e
 export PATH=$XILINX_SDX/bin:$XILINX_VIVADO/bin:$XILINX_SDX/runtime/bin:$PATH
 source "/opt/sdaccel-builder/settings.sh"
-curl -XPOST -H "Content-Type: application/json"  -d '{"status": "STARTED"}' "$CALLBACK_URL" &> /dev/null
+function post_event {
+    curl -XPOST -H "Content-Type: application/json"  -d '{"status": "'"$1"'", "message": "'"$2"'", "code": '${3-0}'}' "$CALLBACK_URL" &> /dev/null
+}
+
+post_event STARTED
 
 set +e
 aws s3 cp --quiet "$INPUT_URL" - | tar zxf - --transform='s,\\,/,g' --show-transformed-names
@@ -10,7 +14,7 @@ aws s3 cp --quiet "$INPUT_URL" - | tar zxf - --transform='s,\\,/,g' --show-trans
 exit="$?"
 
 if [ $exit -ne 0 ]; then
-    curl -XPOST -H "Content-Type: application/json"  -d '{"status": "ERRORED"}' "$CALLBACK_URL" &> /dev/null
+    post_event ERRORED "Source code download failed"
     exit "$exit"
 fi
 
@@ -19,12 +23,21 @@ timeout -k 1m 12h /opt/sdaccel-builder/sdaccel-builder cmds && /opt/sdaccel-buil
 exit="$?"
 
 if [ -n "$DEBUG_URL" ]; then
-    zip -qr artifacts.zip /tmp/workspace/.reco-work
-    aws s3 cp --quiet "artifacts.zip" "$DEBUG_URL"
+    if [ ! -d /tmp/workspace/.reco-work ]; then
+        post_event ERRORED "Cmd compilation failed. Check your host-side code" "$exit"
+        exit "$exit"
+    else
+        zip -qr artifacts.zip /tmp/workspace/.reco-work
+        aws s3 cp --quiet "artifacts.zip" "$DEBUG_URL"
+    fi
 fi
 
 if [ $exit -ne 0 ]; then
-    curl -XPOST -H "Content-Type: application/json"  -d '{"status": "ERRORED"}' "$CALLBACK_URL" &> /dev/null
+    if [ $exit -eq 124 ]; then
+        post_event ERRORED "Build timed out" "$exit"
+    else
+        post_event ERRORED "Unknown error" "$exit"
+    fi
     exit "$exit"
 fi
 
@@ -45,7 +58,7 @@ zip -qr dist.zip .reco-work/sdaccel/dist
 aws s3 cp --quiet "dist.zip" "$OUTPUT_URL"
 
 if [ "$GENERATE_AFI" = "yes" ]; then
-    curl -XPOST -H "Content-Type: application/json"  -d "{\"status\": \"CREATING_IMAGE\", \"message\": \"$AGFI\"}" "$CALLBACK_URL" &> /dev/null
+    post_event CREATING_IMAGE "$AGFI"
 else
-    curl -XPOST -H "Content-Type: application/json"  -d '{"status": "COMPLETED"}' "$CALLBACK_URL" &> /dev/null
+    post_event COMPLETED
 fi