+++ aws/graph.sh
@@ -2,7 +2,12 @@
 set -e
 export PATH=$XILINX_SDX/bin:$XILINX_VIVADO/bin:$XILINX_SDX/runtime/bin:$PATH
 source "/opt/sdaccel-builder/settings.sh"
-curl -XPOST -H "Content-Type: application/json"  -d '{"status": "STARTED"}' "$CALLBACK_URL" &> /dev/null
+
+function post_event {
+    curl -XPOST -H "Content-Type: application/json"  -d '{"status": "'"$1"'", "message": "'"$2"'", "code": '${3-0}'}' "$CALLBACK_URL" &> /dev/null
+}
+
+post_event STARTED
 
 set +e
 aws s3 cp --quiet "$INPUT_URL" - | tar zxf - --transform='s,\\,/,g' --show-transformed-names
@@ -10,7 +15,7 @@ aws s3 cp --quiet "$INPUT_URL" - | tar zxf - --transform='s,\\,/,g' --show-trans
 exit="$?"
 
 if [ $exit -ne 0 ]; then
-    curl -XPOST -H "Content-Type: application/json"  -d '{"status": "ERRORED"}' "$CALLBACK_URL" &> /dev/null
+    post_event ERRORED "Source code download failed"
     exit "$exit"
 fi
 
@@ -19,7 +24,11 @@ timeout -k 1m 30m /opt/sdaccel-builder/sdaccel-builder graph
 exit="$?"
 
 if [ $exit -ne 0 ]; then
-    curl -XPOST -H "Content-Type: application/json"  -d '{"status": "ERRORED"}' "$CALLBACK_URL" &> /dev/null
+    if [ $exit -eq 124 ]; then
+        post_event ERRORED "Graph generation timed out" "$exit"
+    else
+        post_event ERRORED "Unknown error" "$exit"
+    fi
     exit "$exit"
 fi
 
@@ -29,8 +38,8 @@ aws s3 cp --quiet "main-graph.pdf.gz" "$OUTPUT_URL"
 exit="$?"
 
 if [ $exit -ne 0 ]; then
-    curl -XPOST -H "Content-Type: application/json"  -d '{"status": "ERRORED"}' "$CALLBACK_URL" &> /dev/null
+    post_event ERRORED "Graph upload failed" "$exit"
     exit "$exit"
 fi
 
-curl -XPOST -H "Content-Type: application/json"  -d '{"status": "COMPLETED"}' "$CALLBACK_URL" &> /dev/null
+post_event COMPLETED