+++ aws/simulate.sh
@@ -2,14 +2,18 @@
 set -e
 export PATH=$XILINX_SDX/bin:$XILINX_VIVADO/bin:$PATH
 source "/opt/sdaccel-builder/settings.sh"
-curl -XPOST -H "Content-Type: application/json"  -d '{"status": "STARTED"}' "$CALLBACK_URL" &> /dev/null
+function post_event {
+    curl -XPOST -H "Content-Type: application/json"  -d '{"status": "'"$1"'", "message": "'"$2"'", "code": '${3-0}'}' "$CALLBACK_URL" &> /dev/null
+}
+
+post_event STARTED
 
 set +e
 aws s3 cp --quiet "$INPUT_URL" - | tar zxf - --transform='s,\\,/,g' --show-transformed-names
 
 if [ $? -ne 0 ]; then
     exit="$?"
-    curl -XPOST -H "Content-Type: application/json"  -d '{"status": "ERRORED"}' "$CALLBACK_URL" &> /dev/null
+    post_event ERRORED "Source code download failed" "$exit"
     exit "$exit"
 fi
 
@@ -17,13 +21,24 @@ timeout -k 1m 60m /opt/sdaccel-builder/sdaccel-builder simulate "$CMD"
 exit="$?"
 
 if [ $exit -ne 0 ]; then
-    zip -qr artifacts.zip /tmp/workspace/.reco-work
-    aws s3 cp --quiet "artifacts.zip" "$OUTPUT_URL"
+    if [ $exit -eq 124 ]; then
+    	post_event ERRORED "Simulation timed out" "$exit"
+    else
+        if [ -n "$OUTPUT_URL" ]; then
+            if [ ! -d /tmp/workspace/.reco-work ]; then
+                post_event ERRORED "Cmd compilation failed. Check your host-side code" "$exit"
+                exit "$exit"
+            else
+                zip -qr artifacts.zip /tmp/workspace/.reco-work
+                aws s3 cp --quiet "artifacts.zip" "$OUTPUT_URL"
+            fi
+        fi
 
-    curl -XPOST -H "Content-Type: application/json"  -d '{"status": "ERRORED"}' "$CALLBACK_URL" &> /dev/null
+    	post_event ERRORED "Unknown error" "$exit"
+    fi
     exit "$exit"
 fi
 
 cat times.out
 
-curl -XPOST -H "Content-Type: application/json"  -d '{"status": "COMPLETED"}' "$CALLBACK_URL" &> /dev/null
+post_event COMPLETED