+++ go-teak/src/sdaccel/scripts/sda_kernel_build.tcl
@@ -62,6 +62,9 @@
 # -part_family <part_family>
 #   The part family to synthesize for. If not provided, defaults to
 #   "kintexu"
+# -axi_data_width <data_bus_width>
+#   The width of the external AXI data bus to be used for memory access. If
+#   not provided, defaults to 128.
 #
 # The build script can be run from the command line using the Vivado batch mode
 # as follows, where <tcl_script_args> is replaced by the arguments specified
@@ -104,6 +107,7 @@ set paramArgsFileName "param_args.xmldef"
 # Selects a generic Kintex Ultrascale part as the nominal target.
 set partName "xcku115-flvf1924-1-c"
 set partFamily "kintexu"
+set axiDataWidth 128
 
 #
 # Extract the TCL command line arguments.
@@ -165,6 +169,10 @@ while {$argIndex < $argc} {
       set partFamily [lindex $argv $argIndex]
       incr argIndex
     }
+    "-axi_data_width" {
+      set axiDataWidth [lindex $argv $argIndex]
+      incr argIndex
+    }
     default {
       puts "Invalid TCL batch script argument : $arg"
       exit -1
@@ -190,6 +198,15 @@ if {0 == [info exists kernelName]} {
 }
 if {0 == [info exists versionNumber]} {
   puts "Missing VLNV -version argument"
+  exit -1
+}
+
+#
+# Check for valid AXI data bus width.
+#
+if {64 != $axiDataWidth && 128 != $axiDataWidth && \
+  256 != $axiDataWidth && 512 != $axiDataWidth} {
+  puts "Invalid AXI data width : $axiDataWidth"
   exit -1
 }
 
@@ -256,7 +273,8 @@ if {0 != [file exists $constraintFileName]} {
 # Run the standard Xilinx HLS IP packaging flow.
 #
 cd $ipDirPath
-configure_ip_core $moduleName $vendorName $libraryName $kernelName $versionNumber $partFamily
+configure_ip_core $moduleName $vendorName $libraryName $kernelName \
+  $versionNumber $axiDataWidth $partFamily
 cd $buildDirPath
 
 #
@@ -264,4 +282,4 @@ cd $buildDirPath
 # core to create an SDAccel kernel object.
 #
 sda_kernel_packaging $moduleName $vendorName $libraryName $kernelName \
-  $versionNumber $paramArgsFileName $ipDirPath $buildDirPath
+  $versionNumber $axiDataWidth $paramArgsFileName $ipDirPath $buildDirPath